airflow/requirements.txt
geopandas
osm2geojson
google.cloud
